CAMEL-18356: consolidated the Kafka resume strategies
